requirements.txt
aiohttp
botbuilder-core
botbuilder-integration-aiohttp
python-dotenv
langchain
langchain-community
langchain-openai
chromadb
duckdb
pysqlite3-binary
tiktoken
openai



